+++ CHANGES.md
@@ -9,7 +9,10 @@ Version 0.3.2
 - added `bodyHeight` config to `modal` to control modal height
 - added `autocomplete`, `autocorrect`, `autocapitalize` config to `input`
 - added `onMouseDown`, `onMouseUp` events to `button`, `link`
+- added `onTabMenuClick` event to list with `tab` config set to `responsive`
+- added `contains()` to `list`, `fieldset`, `tree`, `table`
 - fixed `getValues()`, `setValues()` work with nested components in `form`
+- fixed responsiveness when `tab` config is set on `list`
 
 Version 0.3.1
 -----------